Update meteoservice.rb
Cities with their corresponding indexes are parsed  from cities.json file now. It is done in order to modify the json file itself and not to tackle with the main code. As well this practice was taught in previous lessons.
Aside of that, changed statement "unless" to "until". With "unless" if the user enters wrong input second time then the program would continue and crash, while with "until" it will prompt user to enter the number again and again until the correct one is matched the condition.